+++ Browser/BrowserWidget.cpp
@@ -216,6 +216,12 @@ void Browser::BrowserWidget::slotLargeIconView()
     changeViewTypeForCurrentView(DB::Category::ThumbedIconView);
 }
 
+void Browser::BrowserWidget::slotSortViewNaturally(bool on)
+{
+    m_filterProxy->setNaturalSortOrder(on);
+    reload();
+}
+
 void Browser::BrowserWidget::changeViewTypeForCurrentView(DB::Category::ViewType type)
 {
     Q_ASSERT(m_list.size() > 0);
@@ -315,7 +321,7 @@ void Browser::BrowserWidget::switchToViewType(DB::Category::ViewType type)
     } else {
         m_curView = m_listView;
         m_filterProxy->invalidate();
-        m_filterProxy->sort(0, Qt::AscendingOrder);
+        m_filterProxy->sort(0, Qt::AscendingOrder); // sort by column 0
 
         m_listView->setViewMode(dynamic_cast<OverviewPage *>(currentAction()) == nullptr ? CenteringIconView::NormalIconView : CenteringIconView::CenterView);
     }