+++ Viewer/ViewerWidget.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2003-2022 The KPhotoAlbum Development Team
+// SPDX-FileCopyrightText: 2003-2023 The KPhotoAlbum Development Team
 //
 // SPDX-License-Identifier: GPL-2.0-or-later
 
@@ -25,6 +25,8 @@ class QMenu;
 class QResizeEvent;
 class QStackedWidget;
 class QWheelEvent;
+class CursorVisibilityHandler;
+class QLabel;
 
 namespace DB
 {
@@ -42,10 +44,12 @@ class InfoDialog;
 }
 namespace Viewer
 {
+
+class AnnotationHandler;
 class AbstractDisplay;
 class ImageDisplay;
 class InfoBox;
-class SpeedDisplay;
+class TransientDisplay;
 class TextDisplay;
 class VideoDisplay;
 class VideoShooter;
@@ -54,11 +58,12 @@ class ViewerWidget : public QStackedWidget
 {
     Q_OBJECT
 public:
-    enum UsageType { InlineViewer,
-                     ViewerWindow };
+    enum class UsageType { CompactPreview, //< Preview ViewerWidget with no (internal) controls. As used in the annotation dialog preview dock.
+                           FullsizePreview, ///< Preview ViewerWidget with minimal context menu that can be used for a full size window. As used in the "fullscreen" preview of tha annotation dialog.
+                           FullFeaturedViewer ///< Full featured ViewerWidget.
+    };
 
-    ViewerWidget(UsageType type = ViewerWindow,
-                 QMap<Qt::Key, QPair<QString, QString>> *macroStore = nullptr);
+    ViewerWidget(UsageType type = UsageType::FullFeaturedViewer);
     ~ViewerWidget() override;
     static ViewerWidget *latest();
     void load(const DB::FileNameList &list, int index = 0);
@@ -121,7 +126,19 @@ protected:
     void createFilterMenu();
     void changeSlideShowInterval(int delta);
     void createVideoViewer();
+    void createAnnotationMenu();
     void inhibitScreenSaver(bool inhibit);
+    /**
+     * @brief currentFileName accesses the current file name in a safe way.
+     * If the current index is invalid, a \c null DB::FileName is returned.
+     * @return the current DB::FileName
+     */
+    DB::FileName currentFileName() const;
+    /**
+     * @brief currentInfo queries the database for the ImageInfo for the current file.
+     * If the current index is invalid, a \c null DB::ImageInfoPtr is returned.
+     * @return the ImageInfoPtr for the current DB::FileName.
+     */
     DB::ImageInfoPtr currentInfo() const;
     friend class InfoBox;
 
@@ -130,11 +147,16 @@ protected:
 private:
     void showNextN(int);
     void showPrevN(int);
-    int find_tag_in_list(const QStringList &list, QString &namefound);
     void invalidateThumbnail() const;
     enum RemoveAction { RemoveImageFromDatabase,
                         OnlyRemoveFromViewer };
     void removeOrDeleteCurrent(RemoveAction);
+
+    enum class TagMode { Locked,
+                         Annotating,
+                         Tokenizing };
+    void setTagMode(TagMode tagMode);
+    void updateContextMenuState(bool isVideo);
 
 protected Q_SLOTS:
     void showNext();
@@ -174,6 +196,9 @@ protected Q_SLOTS:
     void zoomFull();
     void zoomPixelForPixel();
     void makeThumbnailImage();
+    void addTag();
+    void editDescription();
+    void showAnnotationHelp();
 
     /** Set the current window title (filename) and add the given detail */
     void setCaptionWithDetail(const QString &detail);
@@ -185,10 +210,13 @@ protected Q_SLOTS:
     void slotRemoveDeletedImages(const DB::FileNameList &imageList);
 
     void triggerCopyLinkAction(MainWindow::CopyLinkEngine::Action action);
+    void toggleTag(const QString &category, const QString &value);
+    void copyTagsFromPreviousImage();
 
 private:
     static ViewerWidget *s_latest;
     friend class VideoShooter;
+    friend class TemporarilyDisableCursorHandling;
 
     QList<QAction *> m_forwardActions;
     QList<QAction *> m_backwardActions;
@@ -233,12 +261,11 @@ private:
     QAction *m_linkToAction;
 
     InfoBox *m_infoBox;
-    QImage m_currentImage;
 
     bool m_showingFullScreen;
 
     int m_slideShowPause;
-    SpeedDisplay *m_speedDisplay;
+    TransientDisplay *m_transientDisplay;
     KActionCollection *m_actions;
     bool m_forward;
     QTimer *m_slideShowTimer;
@@ -251,20 +278,15 @@ private:
     bool m_videoPlayerStoppedManually;
     UsageType m_type;
 
-    enum InputMode { InACategory,
-                     AlwaysStartWithCategory };
-
-    InputMode m_currentInputMode;
-    QString m_currentInput;
-    QString m_currentCategory;
-    QString m_currentInputList;
-
-    QString m_lastFound;
-    QString m_lastCategory;
-    QMap<Qt::Key, QPair<QString, QString>> *m_inputMacros;
-    QMap<Qt::Key, QPair<QString, QString>> *m_myInputMacros;
-
     MainWindow::CopyLinkEngine *m_copyLinkEngine;
+    CursorVisibilityHandler *m_cursorHandlerForImageDisplay;
+    CursorVisibilityHandler *m_cursorHandlerForVideoDisplay;
+    AnnotationHandler *m_annotationHandler;
+
+    TagMode m_tagMode = TagMode::Locked;
+    QAction *m_addTagAction;
+    QAction *m_copyAction;
+    QAction *m_addDescriptionAction;
 };
 
 }