+++ CHANGELOG.md
@@ -49,6 +49,56 @@ The change log for older releases (before 5.9.0) can be found in CHANGELOG.old.
 
 ### Security
 
+
+KPhotoAlbum 5.12.0 (2023-11-27)
+-------------------------------
+
+### Added
+ - Support annotating images from the viewer by using letters to assign tags.
+   Use the context menu and select "Annotate | Assign Tags" to enable.
+   More information is available in the KPhotoAlbum handbook.
+ - Add option to sort category page by natural order (feature #475339).
+   Natural sort order takes the locale into account and sorts numeric values properly (e.g. sort "9" before "10").
+ - Allow selecting a date range in the DateBar via keyboard (Use "Shift + Left|Right")
+ - Allow closing the annotation dialog's fullscreen preview using the Escape key.
+ 
+### Changed
+ - In the viewer window, using the letters A-Z to assign tokens now needs to be explicitly enabled.
+   You can do this by opening the context menu and selecting "Annotate | Assign Tokens".
+ - When KPhotoAlbum is started in demo mode and a previously saved demo database exists, the old demo database is no longer overwritten.
+ - The ui.rc file (kphotoalbumui.rc) is now deployed as a Qt resource instead of an on-disk file.
+ - Improved usability of "Invoke external program" menu (#474819)
+ - No longer set the default shortcut for "Use current video frame in thumbnail view" to Ctrl+S and avoid shortcut conflict.
+ - Restrict context menu entries for fullscreen preview of annotation dialog to a sane set of actions.
+
+### Dependencies
+ - Add support for exiv2 0.28.1
+ - Remove support for exiv2 < 0.27
+
+### Fixed
+ - Fix issue with KPhotoAlbum not asking to save before exit (#472427)
+ - Fix crash when right-clicking in the empty space of the tag lists of the annotation dialog (#472523)
+ - Prevent showing selected thumbnails only if no thumbnails are selected, so that no crash can occur when showing the context again afterwards (#473324)
+ - Fix crash when switching from video to image (#473587)
+ - Fix program freeze when the viewer window is closed while playing a video using the VLC backend.
+ - Fix crash when both the annotation dialog and the viewer window is open and the user right-clicks on the viewer window (#473762)
+ - Fix several crashes when annotations are not saved and then the user right-clicks on the viewer window (#474151, #476561)
+ - Fix crash when user opens the last image/video in viewer then deletes the image and then accesses the viewer context menu (#474392)
+ - Fix several crashes when deleting an image/video that is currently being annotated in the annotation dialog (#475387, #475388, #476131)
+ - Correctly discard images from annotation dialog if they are deleted elsewhere (e.g. in the thumbnail view).
+ - Fix failed assertion and potential database corruption when searching for new image while the new image search is already running (#475529)
+ - Fix crash when trying to copy or link an image from the annotation dialog's fullscreen preview (#475585)
+ - Use consistent icon sizes in category browser ("Tree" and "Tree with User Icons")
+ - Fix crash when sorting selected images while thumbnail display order is "Newest First" (#476651)
+ - Fix invalid assertion when date bar selection is extended beyond the valid range (#476862)
+ - Fix failed assertion when creating a tag group by drag and drop in a category that does not yet have any tag groups (#477195)
+ - Fix crash when opening a list of files with invalid files (#473231)
+
+### Removed
+ - It is no longer possible to annotate images from the viewer by pressing "/" and typing tag names.
+ - It is no longer possible to change an image through the annotation dialog's fullscreen image preview.
+
+
 KPhotoAlbum 5.11.0 (2023-07-12)
 -------------------------------